Add files via upload
Ok this is the very start of a rpg fiddle. This won't be anything big, it's just something I do to try use some of what I have learned. I am a massive Final Fantasy lover, so it will be slightly inspired by that. For now, only barly scratching the surface of a character creation. But it's fun as heck!